DiscordMessageにメンションされたユーザーの辞書配列を持たせる
