+++ src/domain/entity/command_input.js
@@ -87,6 +87,15 @@ class CommandInput {
     }
 
     /**
+     * 元のDiscordメッセージでメンションされているユーザーの表示名とユーザーIDの辞書配列
+     *
+     * @type {Map<string, string>}
+     */
+    get mentionedUsers() {
+        return this.data.origin.mentionedUsers;
+    }
+
+    /**
      * コマンド引数を１つ消費した新しいエンティティを返す
      *
      * @returns {CommandInput} 1番目の引数が消費されたCommandInput
@@ -123,7 +132,8 @@ class CommandInput {
     }
 
     toString() {
-        return `CommandInput(id=${this.id}, argc=${this.argc}, argv=${this.argv}, serverId=${this.serverId}, channelId=${this.channelId}, voiceChannelId=${this.voiceChannelId})`;
+        const mentionedUsersJson = JSON.stringify(Object.fromEntries(this.mentionedUsers.entries()));
+        return `CommandInput(id=${this.id}, argc=${this.argc}, argv=${this.argv}, serverId=${this.serverId}, channelId=${this.channelId}, voiceChannelId=${this.voiceChannelId}, mentionedUsers=${mentionedUsersJson})`;
     }
 }
