+++ test/unit_specs/domain_model_commands_ask_command_unit_spec.js
@@ -43,6 +43,7 @@ describe('AskCommand', () => {
                     serverId: 'dummy-server-id',
                     channelId: mockTextChannelId,
                     voiceChannelId: null,
+                    mentionedUsers: new Map(),
                 });
                 const input = new CommandInput({
                     id: mockEntityId,
