+++ test/unit_specs/domain_model_commands_seibai_command_unit_spec.js
@@ -1,5 +1,5 @@
 const should = require('chai').should();
-const LeaveCommand = require('../../src/domain/model/commands/leave_command');
+const SeibaiCommand = require('../../src/domain/model/commands/seibai_command');
 const DiscordMessage = require('../../src/domain/entity/discord_message');
 const CommandInput = require('../../src/domain/entity/command_input');
 const ServerStatus = require('../../src/domain/entity/server_status');
@@ -7,15 +7,15 @@ const VoiceStatus = require('../../src/domain/entity/voice_status');
 const Hanako = require('../../src/domain/model/hanako');
 
 /************************************************************************
- * LeaveCommandクラス単体スペック
+ * SeibaiCommandクラス単体スペック
  *
  * メソッド：#process
- * 期待動作：VC退出アクションレスポンスの返却
- * 前提条件：花子VC参加済み
+ * 期待動作：成敗アクションレスポンスの返却
+ * 事前条件：花子VC参加済み ∧ 花子音声配信中
  * 備考：なし
  ***********************************************************************/
 
-describe('LeaveCommand', () => {
+describe('SeibaiCommand', () => {
     // helper
     function serverStatusBlueprint() {
         return {
@@ -45,24 +45,24 @@ describe('LeaveCommand', () => {
 
     describe('Commandクラスメタスペック', () => {
         specify('typeは文字列を返す', () => {
-            const sub = new LeaveCommand(basicHanako());
+            const sub = new SeibaiCommand(basicHanako());
             sub.type.should.be.a('string');
         });
 
         specify('namesは静的に文字列の配列を返す', () => {
-            LeaveCommand.names.should.be.an('array').that.is.not.empty;
-            LeaveCommand.names.forEach(name => name.should.be.a('string'));
+            SeibaiCommand.names.should.be.an('array').that.is.not.empty;
+            SeibaiCommand.names.forEach(name => name.should.be.a('string'));
         });
 
         specify('processメソッドを持つ', () => {
-            const sub = new LeaveCommand(basicHanako());
+            const sub = new SeibaiCommand(basicHanako());
             sub.process.should.be.a('function');
         });
     });
 
     describe('#process', () => {
         context('正常系', () => {
-            specify('正しいVC退出アクションレスポンスを返す', () => {
+            specify('正しい成敗アクションレスポンスを返す', () => {
                 const mockEntityId = 'mock-001';
                 const mockServerId = 'mock-server-001';
                 const mockVoiceChannelId = 'mock-voice-001';
@@ -75,6 +75,7 @@ describe('LeaveCommand', () => {
                     serverId: mockServerId,
                     channelId: mockTextChannelId,
                     voiceChannelId: mockVoiceChannelId,
+                    mentionedUsers: new Map(),
                 });
                 const input = new CommandInput({
                     id: mockEntityId,
@@ -88,18 +89,21 @@ describe('LeaveCommand', () => {
 
                 const vsb = voiceStatusBlueprint();
                 vsb.serverId = mockServerId;
+                vsb.state = 'speaking';
                 vsb.voiceChannelId = mockVoiceChannelId;
                 vsb.readingChannelsId = [mockTextChannelId];
 
-                const sub = new LeaveCommand(new Hanako(new ServerStatus(ssb), new VoiceStatus(vsb), null));
+                const sub = new SeibaiCommand(new Hanako(new ServerStatus(ssb), new VoiceStatus(vsb), null));
                 const res = sub.process(input);
 
                 // 正しいアクションレスポンス
                 res.type.should.equal('action');
                 res.id.should.equal(mockEntityId);
-                res.action.type.should.equal('leave_voice');
+                res.action.type.should.equal('seibai');
                 res.action.serverId.should.equal(mockServerId);
-                res.onSuccess.type.should.equal('silent');
+                res.onSuccess.type.should.equal('chat');
+                res.onSuccess.code.should.equal('simple');
+                res.onSuccess.channelId.should.equal(mockTextChannelId);
                 res.onFailure.type.should.equal('silent');
             });
         });
@@ -114,6 +118,7 @@ describe('LeaveCommand', () => {
                         serverId: 'mockserver',
                         channelId: 'mockchannel',
                         voiceChannelId: 'mockchannel2',
+                        mentionedUsers: new Map(),
                     });
                     const input = new CommandInput({
                         id: 'mock',
@@ -124,7 +129,39 @@ describe('LeaveCommand', () => {
 
                     const ssb = serverStatusBlueprint();
 
-                    const sub = new LeaveCommand(new Hanako(new ServerStatus(ssb), null, null));
+                    const sub = new SeibaiCommand(new Hanako(new ServerStatus(ssb), null, null));
+                    const res = sub.process(input);
+
+                    // エラー会話レスポンス
+                    res.type.should.equal('chat');
+                    res.code.should.equal('error');
+                });
+            });
+
+            context('花子はVCに参加しているが、音声配信中ではない', () => {
+                specify('エラー会話レスポンスを返す', () => {
+                    const dmessage = new DiscordMessage({
+                        id: 'mock',
+                        content: 'dummy',
+                        type: 'command',
+                        serverId: 'mockserver',
+                        channelId: 'mockchannel',
+                        voiceChannelId: 'mockchannel2',
+                        mentionedUsers: new Map(),
+                    });
+                    const input = new CommandInput({
+                        id: 'mock',
+                        argc: 0,
+                        argv: [],
+                        origin: dmessage,
+                    });
+
+                    const ssb = serverStatusBlueprint();
+
+                    const vsb = voiceStatusBlueprint();
+                    vsb.state = 'ready';
+
+                    const sub = new SeibaiCommand(new Hanako(new ServerStatus(ssb), new VoiceStatus(vsb), null));
                     const res = sub.process(input);
 
                     // エラー会話レスポンス
